+++ docs/src/components/Link/Link.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { Link as GatsbyLink, GatsbyLinkProps } from 'gatsby';
+import { Link as MarigoldLink } from '@marigold/components';
+
+export const Link: React.FC<GatsbyLinkProps<unknown>> = ({
+  children,
+  ...props
+}) => <MarigoldLink as={GatsbyLink}>{children}</MarigoldLink>;
